Microservice/deploy-management-server/template_man/requirements.txt
fastapi==0.111.1
pyyaml==6.0.1
uvicorn==0.30.3
# kubernetes==30.1.0